+++ CHANGELOG.md
@@ -5,10 +5,13 @@
 - Improved `plotDEheatmap` function
 - Fixed bug with `balancing.factor.per.sample` in `buildGraph`
 - Fixed some installation problems
+- Changed `std::cout` to `Rcpp::Rcout` (July 2020)
+- Revised README, vignettes (July 2020)
 
-### Added
+### Added 
 
 - Added checks for `getPerCellTypeDE()` for errors, removing NAs
+- Added `ht_opt$message = FALSE` for ComplexHeatmap (July 2020)
 - Write to HDF5 for `saveConosForScanPy()` (7 July 2020)
 - LICENSE (July 2020)